requirements.txt
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.1
Pillow>=8.2.0
numpy>=1.19.2